+++ Languages/JavaScript/Node/Tutorial/node-http/index.cjs
@@ -1,6 +1,6 @@
 const http = require('node:http') // HTTP - Hypertext Transfer Protocol
 const fs = require('node:fs')
-const { findAvailablePort } = require('@fjrodafo/free-port')
+const { findAvailablePort } = require('@fjrodafo/port-finder')
 
 // Status Code https://http.cat/